test(core) astubbs#121: one racing double, and a guard that the seam actually fired
The two reproductions carried near-identical test doubles - same
getOffsetToCommit override, same read ledger, same firstOffsetToCommitRead, and a
snapshot override differing only in one-shot vs re-armable. The compounding
class's javadoc defended the split as "state the sibling does not need", which
does not hold: arm once and never re-arm IS the one-shot behaviour, so the
re-armable version is a strict superset. Extracted to one package-private class
and deleted both copies.

Its javadoc no longer claims what it cannot. The old text said a completion after
the snapshot "can change neither the payload nor offsetHighestSucceeded". That is
true of these fixtures, because every offset they race is the lowest outstanding
one and therefore already below the high-water mark - not of the seam.
encodeOffsetsCompressed takes its own second read of getOffsetHighestSucceeded()
after the snapshot, so a completion ABOVE the mark would widen the encoder's range
against a stale set. These tests do not reach that interleaving and now say so.

Added the guard the review asked for: an assertion that the injected race actually
fired. Without it, both classes pass on the fixed code for a trivial reason - both
numbers come from one sample - and would keep passing if the seam silently stopped
being called.

The first version of that guard was itself tautological, and mutation-checking is
what caught it. It inferred firing from `armedRacingOffset == null`, which cannot
distinguish "armed, then fired" from "never armed": deleting the arm left the
suite green. Firing is now tracked explicitly, and removing the arm turns the
precondition red.

Verified in both directions after the refactor, because a deduplication that
quietly disarms the tests would look exactly like a successful one: 7 of 7 red
against master's unfixed PartitionState, 7 of 7 green against the fix.

Refs: confluentinc#894, astubbs#121
