+++ parallel-consumer-core/src/test/java/bz/stub/parallelconsumer/state/PartitionStateCommitShiftCompounding894Test.java
@@ -464,64 +464,4 @@ class PartitionStateCommitShiftCompounding894Test {
         return OffsetMapCodecManager.deserialiseIncompleteOffsetMapFromBase64(committed.offset(), committed.metadata());
     }
 
-    /**
-     * A {@link PartitionState} that lands one work completion inside a commit cycle, at the moment the encoder has
-     * finished reading the state it encodes, and can be re-armed for the next cycle.
-     * <p>
-     * Same seam and same reasoning as the sibling test's racing state - {@code encodeOffsetsCompressed} calls
-     * {@link #getIncompleteOffsetsBelowHighestSucceeded()} exactly once to snapshot its input, and the snapshot is a
-     * fresh copy, so a completion landing immediately after it can change neither the payload nor
-     * {@code offsetHighestSucceeded}. It is a separate class rather than a shared helper because it carries state
-     * the sibling does not need: an offset that varies per cycle and a re-arm, both required to run the cycle more
-     * than once. Merging the two would push per-cycle machinery into a test that has no cycles.
-     */
-    private static class RacingCommitCycleState extends PartitionState<String, String> {
-
-        private final List<Long> offsetToCommitReads = new ArrayList<>();
-
-        /** The offset to complete when the encoder next takes its snapshot, or {@code null} when not armed. */
-        private Long armedRacingOffset;
-
-        RacingCommitCycleState(PCModule<String, String> module,
-                               TopicPartition topicPartition,
-                               HighestOffsetAndIncompletes offsetData) {
-            super(0, module, topicPartition, offsetData);
-        }
-
-        void armRaceOn(long offset) {
-            this.armedRacingOffset = offset;
-        }
-
-        @Override
-        protected long getOffsetToCommit() {
-            long read = super.getOffsetToCommit();
-            offsetToCommitReads.add(read);
-            return read;
-        }
-
-        @Override
-        public SortedSet<Long> getIncompleteOffsetsBelowHighestSucceeded() {
-            SortedSet<Long> snapshotTheEncoderWillUse = super.getIncompleteOffsetsBelowHighestSucceeded();
-            if (armedRacingOffset != null) {
-                long racing = armedRacingOffset;
-                armedRacingOffset = null;
-                log.debug("Racing completion of offset {} lands after the encoder snapshot {}",
-                        racing, snapshotTheEncoderWillUse);
-                onSuccess(racing);
-            }
-            return snapshotTheEncoderWillUse;
-        }
-
-        /**
-         * @return the read that {@code tryToEncodeOffsets} used as the payload's base - the first on the unfixed
-         *         code, which reads twice, and the only one on the fixed code, which reads once and threads it
-         *         through a tuple
-         */
-        long firstOffsetToCommitRead() {
-            assertWithMessage("the encode path must have read the offset to commit at least once")
-                    .that(offsetToCommitReads)
-                    .isNotEmpty();
-            return offsetToCommitReads.get(0);
-        }
-    }
 }
